chore: cleaned up loops where initialisers did not match the datatypes of the comparitors chore: replaced = initialisations with new initialiser syntax

+++ source/includes/integration/finite_volumes/legacy_transport_scheme/StencilProcessor.h
@@ -116,7 +116,7 @@ struct StencilProcessor {
     void CorrectImplicitTwoPhaseSolutionAtBoundary_NonlinearNewtonRaphson(const FV_Parameter& param,
                                               const Element<dim>& e,
                                               TwoPhaseModel<dim>& relperm,
-                                              size_t pnid,
+                                              uint32_t pnid,
                                               double& flux,
                                               bool with_gravity_forces,
                                               bool with_capillary_spreading);
